Fixed all outposts and Airfields
All outposts have had more Emplacement spawns added to them and have been optimized so that useless props aren't present anymore and for the objects that are there their simulation has been disabled. Airfields now also have helipads which will have helicopters spawning on them. The height on most of the outposts should be fixed too as well as some markers that shouldn't be visible.